# .agents/skills/hashtagweb3-agent-skill/public/openapi.yaml
openapi: 3.1.0
info:
  title: Hashtag Web3 Public API
  version: 1.0.0
  description: Public REST API for Hashtag Web3 (hashtagweb3.com). Search live Web3 jobs, crypto industry news, global blockchain conferences & hackathons, and the 200+ term Web3 glossary. Optimized for LLM agents, automated scripts, and developer integrations.
  x-api-versioning:
    policy: "URI path versioning (/api/v1/)"
    deprecation: "We signal deprecation using standard Deprecation and Sunset HTTP headers in API responses, with a minimum 12-month advance notice period."
    policyUrl: https://hashtagweb3.com/api-policy
    documentationUrl: https://hashtagweb3.com/developers
  contact:
    name: Hashtag Web3 Developer Support
    url: https://hashtagweb3.com/developers
    email: contact@hashtagweb3.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://hashtagweb3.com/api/v1
    description: Production server (v1)
  - url: https://hashtagweb3.com/api/sandbox
    description: Sandbox / Test Environment (Mock Data, Zero-Auth, Safe Testing)
  - url: https://hashtagweb3.com
    description: Production server root
paths:
...
  /api/news:
    get:
      operationId: listNews
      summary: List Web3 & Crypto Industry News
      description: Retrieves real-time aggregated crypto and blockchain news headlines from premier industry publications.
      parameters:
        - name: search
          in: query
          required: false
          description: Filter news headlines by keyword.
          schema:
            type: string
            example: Ethereum
        - name: limit
          in: query
          required: false
          description: Maximum number of news articles to return (default 30, max 100).
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 100
            example: 15
      responses:
        '200':
          description: List of recent crypto news headlines.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsListResponse'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
components:
  schemas:
...
    NewsItem:
      type: object
      required: [title, link, pubDate, source]
      properties:
        title:
          type: string
          example: Ethereum Layer 2 Activity Reaches Record High
        link:
          type: string
          format: uri
          example: https://cointelegraph.com/...
        pubDate:
          type: string
          format: date-time
          example: "2026-08-24T14:30:00Z"
        source:
          type: string
          example: Cointelegraph
        contentSnippet:
          type: string
          example: Daily active addresses across major rollups surged...
    NewsListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/NewsItem'
        meta:
          type: object
          required: [total, count]
          properties:
            total:
              type: integer
              example: 50
            count:
              type: integer
              example: 30
...
    ApiErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, hint]
          properties:
            code:
              type: string
              enum: [BAD_REQUEST, NOT_FOUND, INTERNAL_SERVER_ERROR, RATE_LIMITED]
              example: BAD_REQUEST
            message:
              type: string
              example: "Invalid parameter: 'limit' must be an integer between 1 and 200."
            hint:
              type: string
              example: "Adjust the limit parameter and retry. View docs at https://hashtagweb3.com/developers"
            docUrl:
              type: string
              format: uri
              example: https://hashtagweb3.com/developers